common.py: read all data in readCve() and return dict()
Before readCveHeaders would return an EmailMessage and wouldn't parse
the 'content' of the message. This 'content' in the context of a CVE is
actually just more headers, so feed that back into HeaderParser, then
put all of it in a dict() instead of an EmailMessage. This keeps the
file reading format in common.py and we can use standard dictionary
everywhere else.