cvelib/common.py: allow uppercase in software name for non-compat mode
GitHub allows repos to have uppercase letters so we should allow the
software name to have them too. For now, only apply this to non-compat
mode since in Ubuntu they currently only use Debian package names which
are not allowed to use uppercase.

A future improvement might be to allow uppercase only with github.