add readCacheDB: TODO tests
Note: this is not significantly faster. Reading 400+ files off the disk
and parsing them is under .8 seconds whereas having them all in a JSON
file (these commits) with no parsing/verifying is under .7 seconds.

Creating an sqlite db as a cache could be useful for certain types of
queries where a we want arbitrary SQL against a relational model. That
can be explored in the future.

This branch might be interesting for an easier machine-readable export,
but there is no demand for that at present.

If pursuing this, might do:
  * cve-query/report tool reads in files and outputs Json/SQLite as cache. If
    cache exists, check mtime of it against highest mtime of all files. If
    lower, regenerate, else read in (this is what is implemented now for
    cve-report* via collectData)
  * check-syntax can output cache on a successful run. Next times, find mtimes
    of newer files than cache and only check those
  * Cache:
    ```
    {
     "version": 1,
     "cves": (
       "CVE-...": {},
     },
     "software": { "<software>:": ["CVE-...", ...]},
     "gh-dependabot": { "<dep>": ["CVE-...", ....]},
     ... search indexes ...
    }
    ```
    'version' and 'cves' is implemented in this branch now